backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
supabase==2.3.0
asyncpg==0.29.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
openai==1.3.0
httpx==0.24.1
dnspython==2.6.1
email-validator==2.1.0.post1